refactor: move the synthon package to the top level
Enumeration was one of thirteen modules, so naming the package for it described an eighth of its contents while fragment.py, coverage.py and priority.py sat under a word that contradicts them.
