+++ docs/api.rst
@@ -15,7 +15,7 @@ The following subpackages are public.
 - ``synplan.mcts``: Functions and classes responsible for Monte-Carlo Tree Search.
 - ``synplan.ml``: Functions that are used to train policy and value networks.
 - ``synplan.chem.reaction.routes``: Post-search route processing, including RouteCGR construction, route IO, clustering, route analysis, depiction, and quality scoring. The route quality scoring module is inspired by Westerlund et al., *ChemRxiv*, 2025 (`doi:10.26434/chemrxiv-2025-gdrr8 <https://doi.org/10.26434/chemrxiv-2025-gdrr8>`_).
-- ``synplan.enumeration.synthon``: Adapted Synt-On workflows for building-block classification, synthonisation, fragmentation, enumeration, stock analysis, and MCTS priority rules.
+- ``synplan.synthon``: Adapted Synt-On workflows for building-block classification, synthonisation, fragmentation, enumeration, stock analysis, and MCTS priority rules.
 - ``synplan.utils``: Functions used for configuring ``SynPlanner``, loading all the data, logging and visualisation of synthetic routes.
 
 .. toctree::
