+++ docs/api_reference/synplan.synthon.rst
@@ -1,20 +1,20 @@
-synplan.enumeration.synthon package
-===================================
+synplan.synthon package
+=======================
 
-``synplan.enumeration.synthon`` provides the adapted Synt-On workflows for
+``synplan.synthon`` provides the adapted Synt-On workflows for
 building-block classification and synthonisation, target fragmentation,
 analogue-aware enumeration, stock handling, scaffold analysis, and MCTS
 priority rules.
 
 Use the package-level imports for the stable public facade, for example
-``from synplan.enumeration.synthon import Fragmenter, SynthonConfig``. Focused
+``from synplan.synthon import Fragmenter, SynthonConfig``. Focused
 submodules expose advanced APIs and identify where each implementation is
 owned.
 
 Public package API
 ------------------
 
-.. automodule:: synplan.enumeration.synthon
+.. automodule:: synplan.synthon
    :members:
    :undoc-members:
    :show-inheritance:
@@ -22,98 +22,98 @@ Public package API
 Submodules
 ----------
 
-synplan.enumeration.synthon.analogues module
+synplan.synthon.analogues module
 -------------------------------------
 
-.. automodule:: synplan.enumeration.synthon.analogues
+.. automodule:: synplan.synthon.analogues
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.audit module
+synplan.synthon.audit module
 ---------------------------------
 
-.. automodule:: synplan.enumeration.synthon.audit
+.. automodule:: synplan.synthon.audit
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.classify module
+synplan.synthon.classify module
 ------------------------------------
 
-.. automodule:: synplan.enumeration.synthon.classify
+.. automodule:: synplan.synthon.classify
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.cli module
+synplan.synthon.cli module
 -------------------------------
 
-.. automodule:: synplan.enumeration.synthon.cli
+.. automodule:: synplan.synthon.cli
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.config module
+synplan.synthon.config module
 ----------------------------------
 
-.. automodule:: synplan.enumeration.synthon.config
+.. automodule:: synplan.synthon.config
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.enumeration module
+synplan.synthon.enumeration module
 ---------------------------------------
 
-.. automodule:: synplan.enumeration.synthon.enumeration
+.. automodule:: synplan.synthon.enumeration
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.fragment module
+synplan.synthon.fragment module
 ------------------------------------
 
-.. automodule:: synplan.enumeration.synthon.fragment
+.. automodule:: synplan.synthon.fragment
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.priority module
+synplan.synthon.priority module
 ------------------------------------
 
-.. automodule:: synplan.enumeration.synthon.priority
+.. automodule:: synplan.synthon.priority
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.reactor module
+synplan.synthon.reactor module
 -----------------------------------
 
-.. automodule:: synplan.enumeration.synthon.reactor
+.. automodule:: synplan.synthon.reactor
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.scaffolds module
+synplan.synthon.scaffolds module
 -------------------------------------
 
-.. automodule:: synplan.enumeration.synthon.scaffolds
+.. automodule:: synplan.synthon.scaffolds
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.stock module
+synplan.synthon.stock module
 ---------------------------------
 
-.. automodule:: synplan.enumeration.synthon.stock
+.. automodule:: synplan.synthon.stock
    :members:
    :undoc-members:
    :show-inheritance:
 
-synplan.enumeration.synthon.synthonise module
+synplan.synthon.synthonise module
 --------------------------------------
 
-.. automodule:: synplan.enumeration.synthon.synthonise
+.. automodule:: synplan.synthon.synthonise
    :members:
    :undoc-members:
    :show-inheritance:
