+++ synplan/synthon/classify.py
@@ -4,7 +4,7 @@ from chython import smarts, smiles
 from chython.containers import MoleculeContainer, QueryContainer
 
 from synplan.chem.utils import safe_canonicalization
-from synplan.enumeration.synthon.config import SynthonConfig, load_data
+from synplan.synthon.config import SynthonConfig, load_data
 
 
 class SynthonDataError(ValueError):
