+++ tests/regression/test_synthon_reaction_coverage.py
@@ -10,15 +10,15 @@ from pathlib import Path
 import pytest
 from chython import smarts, smiles
 
-from synplan.enumeration.synthon import cli
-from synplan.enumeration.synthon.config import SynthonConfig, load_data
-from synplan.enumeration.synthon.coverage import (
+from synplan.synthon import cli
+from synplan.synthon.config import SynthonConfig, load_data
+from synplan.synthon.coverage import (
     CoverageRule,
     classify_coverage,
     load_coverage_rules,
 )
-from synplan.enumeration.synthon.priority import capped_smarts
-from synplan.enumeration.synthon.reactor import (
+from synplan.synthon.priority import capped_smarts
+from synplan.synthon.reactor import (
     RULE_NUCLEOPHILE_CAPS,
     SynthonRuleError,
     query_labels,
@@ -100,7 +100,7 @@ def test_a_rule_that_lost_its_labels_is_refused():
 
 def test_the_named_nucleophiles_come_from_the_capping_table():
     """One table, not two: the rules that must SPELL a reagent are the rules that must SEE it."""
-    from synplan.enumeration.synthon.coverage import _RULE_NUCLEOPHILE_ELEMENTS
+    from synplan.synthon.coverage import _RULE_NUCLEOPHILE_ELEMENTS
 
     assert set(_RULE_NUCLEOPHILE_ELEMENTS) == set(RULE_NUCLEOPHILE_CAPS)
     assert _RULE_NUCLEOPHILE_ELEMENTS["R12.3a"] == frozenset(("B",))
