+++ tests/unit/chem/synthon/test_classify.py
@@ -6,12 +6,12 @@ import pytest
 from chython import smiles
 
 from synplan.chem.utils import safe_canonicalization
-from synplan.enumeration.synthon.classify import (
+from synplan.synthon.classify import (
     BBClassifier,
     SynthonDataError,
     _compile,
 )
-from synplan.enumeration.synthon.config import SynthonConfig, load_data
+from synplan.synthon.config import SynthonConfig, load_data
 
 FIXTURES = Path(__file__).resolve().parents[3] / "data" / "synthon"
