+++ tests/unit/chem/synthon/test_stock.py
@@ -4,8 +4,8 @@ import pytest
 from chython import synthon_smiles
 from pydantic import ValidationError
 
-from synplan.enumeration.synthon.config import SynthonConfig
-from synplan.enumeration.synthon.stock import (
+from synplan.synthon.config import SynthonConfig
+from synplan.synthon.stock import (
     SynthonRecord,
     SynthonStock,
     cap_leaving_group,
